// out/model/userCreate.ts
/**
 * TransferTo Calculator API
 * TransferTo Calculator
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { User } from './user';


export interface UserCreate {
    /**
     * Middle name
     */
    middle_name?: string;
    /**
     * The email used for authentication
     */
    email: string;
    /**
     * Last name
     */
    last_name?: string;
    /**
     * First name
     */
    first_name?: string;
    /**
     * The password used for authentication
     */
    password?: string;
}